solving out of memory error on brain_large dataset

+++ scvi/train.py
@@ -16,7 +16,7 @@ def train(
             i_batch,
             (sample_batch, local_l_mean, local_l_var, batch_index, _),
         ) in enumerate(data_loader_train):
-
+            sample_batch = sample_batch.type(torch.FloatTensor)
             sample_batch = Variable(sample_batch)
             local_l_mean = Variable(local_l_mean)
             local_l_var = Variable(local_l_var)
